refactor: extract cluster-stats logic out of app.py (break-up phase 1)
Move the ClusterStats/PendingPartitionStats data models and the pure
cluster-stats computation off the SlurmMonitor god-object into a new
dependency-free domain module stoei/cluster_stats.py (imports only
stoei.slurm.* + stdlib). cluster_sidebar.py re-exports the models so
existing importers resolve. SlurmMonitor keeps thin delegating methods,
so every call site and test is unchanged.

app.py: 2968 -> 2771 lines.
